Uso del nuevo Gridbox formateado los campos de metodo index corregidos algunos errores logicos Ahora la existencia es un decimal

+++ database/migrations/2021_12_02_141238_create_primaries_products_table.php
@@ -16,7 +16,7 @@ class CreatePrimariesProductsTable extends Migration
         Schema::create('primaries_products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('stock');
+            $table->decimal('stock', 10, 2);
             $table->timestamps();
         });
     }
